+++ code/crates/sync/src/state.rs
@@ -69,7 +69,7 @@ where
 
     /// The maximum number of parallel requests that can be made to peers.
     /// If the configuration is set to 0, it defaults to 1.
-    pub fn max_parallel_requests(&self) -> u64 {
+    pub fn max_parallel_requests(&self) -> usize {
         max(1, self.config.parallel_requests)
     }
